ref: create dedicated parser module and update conftests
