+++ tests/io/qchem/conftest.py
@@ -7,39 +7,45 @@ from calcflow.io.qchem import parse_qchem_sp
 
 
 @pytest.fixture(scope="session")
-def qchem_h2o_62_sp_output(testing_data_path: Path) -> str:
-    """Fixture that reads the content of the QChem H2O SP output file."""
-    file_path = testing_data_path / "qchem" / "h2o" / "6.2-sp-smd.out"
-    return file_path.read_text()
+def parsed_qchem_54_h2o_sp_data(testing_data_path: Path) -> CalculationResult:
+    return parse_qchem_sp((testing_data_path / "qchem" / "h2o" / "5.4-sp-smd.out").read_text())
 
 
 @pytest.fixture(scope="session")
-def parsed_qchem_62_h2o_sp_data(qchem_h2o_62_sp_output: str) -> CalculationResult:
-    """Fixture that provides the fully parsed CalculationResult object."""
-    return parse_qchem_sp(qchem_h2o_62_sp_output)
+def parsed_qchem_62_h2o_sp_data(testing_data_path: Path) -> CalculationResult:
+    return parse_qchem_sp((testing_data_path / "qchem" / "h2o" / "6.2-sp-smd.out").read_text())
 
 
 @pytest.fixture(scope="session")
-def qchem_h2o_54_sp_output(testing_data_path: Path) -> str:
-    """Fixture that reads the content of the QChem H2O SP output file."""
-    file_path = testing_data_path / "qchem" / "h2o" / "5.4-sp-smd.out"
-    return file_path.read_text()
+def parsed_qchem_54_h2o_uks_tddft_data(testing_data_path: Path) -> CalculationResult:
+    return parse_qchem_sp((testing_data_path / "qchem" / "h2o" / "5.4-uks-tddft.out").read_text())
 
 
 @pytest.fixture(scope="session")
-def parsed_qchem_54_h2o_sp_data(qchem_h2o_54_sp_output: str) -> CalculationResult:
-    """Fixture that provides the fully parsed CalculationResult object."""
-    return parse_qchem_sp(qchem_h2o_54_sp_output)
+def parsed_qchem_62_h2o_uks_tddft_data(testing_data_path: Path) -> CalculationResult:
+    return parse_qchem_sp((testing_data_path / "qchem" / "h2o" / "6.2-uks-tddft.out").read_text())
 
 
 @pytest.fixture(scope="session")
-def qchem_h2o_62_uks_tddft_output(testing_data_path: Path) -> str:
-    """Fixture that reads the content of the QChem H2O SP output file."""
-    file_path = testing_data_path / "qchem" / "h2o" / "6.2-uks-tddft.out"
-    return file_path.read_text()
+def parsed_qchem_62_h2o_rks_tddft_data(testing_data_path: Path) -> CalculationResult:
+    return parse_qchem_sp((testing_data_path / "qchem" / "h2o" / "6.2-rks-tddft.out").read_text())
 
 
 @pytest.fixture(scope="session")
-def parsed_qchem_62_h2o_uks_tddft_data(qchem_h2o_62_uks_tddft_output: str) -> CalculationResult:
-    """Fixture that provides the fully parsed CalculationResult object."""
-    return parse_qchem_sp(qchem_h2o_62_uks_tddft_output)
+def parsed_qchem_54_h2o_mom_sp_data(testing_data_path: Path) -> CalculationResult:
+    return parse_qchem_sp((testing_data_path / "qchem" / "h2o" / "5.4-mom-sp-smd.out").read_text())
+
+
+@pytest.fixture(scope="session")
+def parsed_qchem_62_h2o_mom_sp_data(testing_data_path: Path) -> CalculationResult:
+    return parse_qchem_sp((testing_data_path / "qchem" / "h2o" / "6.2-mom-sp-smd.out").read_text())
+
+
+@pytest.fixture(scope="session")
+def parsed_qchem_54_h2o_mom_xas_data(testing_data_path: Path) -> CalculationResult:
+    return parse_qchem_sp((testing_data_path / "qchem" / "h2o" / "5.4-mom-xas-smd.out").read_text())
+
+
+@pytest.fixture(scope="session")
+def parsed_qchem_62_h2o_mom_xas_data(testing_data_path: Path) -> CalculationResult:
+    return parse_qchem_sp((testing_data_path / "qchem" / "h2o" / "6.2-mom-xas-smd.out").read_text())
